requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Graph processing
networkx>=3.0
igraph>=0.10.0

# Dataset handling
ogb>=1.3.6
pandas>=2.0.0

# Visualization and logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

